iadevideosmark/apps/api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-multipart>=0.0.9
httpx>=0.27
python-dotenv>=1.0
Pillow>=10.0
aiofiles>=23.0
